pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = ""
authors = ["christianmasa <christian@linkernetworks.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
drf-spectacular = "^0.22.1"
pre-commit = "^2.18.1"
django-extensions = "^3.1.5"
drf-flex-fields = "^0.9.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
